+++ src/test/java/engine/model/physicalEngine/shape/RectangleTest.java
@@ -1,64 +1,73 @@
 package engine.model.physicalEngine.shape;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import engine.model.physicalEngine.shape.temp.Position;
-import org.junit.jupiter.api.Test;
+import engine.model.physicalEngine.movement.*;
 
 public class RectangleTest {
     @Test
-    void testgetHeight(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        assertEquals(rec.getWidth(),5);
+    void testgetHeight() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        assertEquals(rec.getWidth(), 5);
     }
+
     @Test
-    void testsetHeight(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        double a=10;
+    void testsetHeight() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        float a = 10;
         rec.setWidth(a);
-        assertEquals(rec.getWidth(),a);
+        assertEquals(rec.getWidth(), a);
     }
+
     @Test
-    void testgetLength(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        assertEquals(rec.getLength(),10);
+    void testgetLength() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        assertEquals(rec.getLength(), 10);
     }
+
     @Test
-    void testsetLength(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        double a=10;
+    void testsetLength() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        float a = 10;
         rec.setLength(a);
-        assertEquals(rec.getLength(),a);
+        assertEquals(rec.getLength(), a);
     }
+
     @Test
-    void testsetCenter(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        rec.setCenter(new Position(5,5));
-        assertEquals(rec.getCenter().getX(),5);
-        assertEquals(rec.getCenter().getY(),5);
+    void testsetCenter() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        rec.setCenter(new Position(5, 5));
+        assertEquals(rec.getCenter().getX(), 5);
+        assertEquals(rec.getCenter().getY(), 5);
     }
+
     @Test
-    void testgetCenter(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        assertEquals(rec.getCenter().getX(),1);
-        assertEquals(rec.getCenter().getY(),1);
+    void testgetCenter() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        assertEquals(rec.getCenter().getX(), 1);
+        assertEquals(rec.getCenter().getY(), 1);
     }
+
     @Test
-    void testisMoving(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        assertEquals(rec.isMoving(),false);
+    void testisMoving() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        assertEquals(rec.isMoving(), false);
     }
+
     @Test
-    void testsetIsMoving(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        boolean bool=true;
+    void testsetIsMoving() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        boolean bool = true;
         rec.setIsMoving(bool);
-        assertEquals(rec.isMoving(),true);
+        assertEquals(rec.isMoving(), true);
     }
+
     @Test
-    void testgetApex(){
-        Rectangle rec=new Rectangle(1,1,10,5);
-        List<Position> list= rec.getApex();
+    void testgetApex() {
+        Rectangle rec = new Rectangle(1, 1, 10, 5);
+        List<Position> list = rec.getApex();
 
         assertEquals(list.get(0).getX(), -4);
         assertEquals(list.get(0).getY(), -1.5);
